backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
openai==1.3.0
httpx==0.25.0
python-dotenv==1.0.0
requests==2.31.0